Update D8 R8 master
Version: master 5a8c428c672719160ac5b2aa5cdfc892ac045500
This is a development snapshot, it's fine to use for studio
canary build, but not for BETA or release, for those we
would need a release version of R8 binaries. This build
IS suitable for preview release but IS NOT suitable for
public release.

Built here:
go/r8-releases/raw/master/5a8c428c672719160ac5b2aa5cdfc892ac045500/
Test: ./gradlew check

Change-Id: I29b54f367c42f8577fd9bfce2233096b1a6a85c6