Update D8 R8 master
Version: master 7a7c49accc59352276816718a4c7ae52f9683f99
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/7a7c49accc59352276816718a4c7ae52f9683f99/
Test: ./gradlew check
Bug:

Change-Id: Iec2d3d51e22ab9fa3de484d2e5e65b29eccbfb9c