Update D8 R8 master
Version: master 79c10fe71304b74d92b41313287249f56b6f2134
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/79c10fe71304b74d92b41313287249f56b6f2134/
Test: ./gradlew check
Bug:

Change-Id: If082e42d217a87c5340e64dc88fd132513c3cc48