Update D8 R8 master to 1.2.21-dev
Version: master 48962963178f8a971a5a536310d2556b457a3bfe
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for
public release.

Built here: go/r8-releases/raw/master/48962963178f8a971a5a536310d2556b457a3bfe/
Test: ./gradlew check
Bug:

Change-Id: If65586268a502a1ad03c5f8ab89c09ca1eb0f790