Update D8 R8 master to 1.3.4-dev
Version: master b94cbceddde92651c5a75cd5195768ed34d54be7
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/b94cbceddde92651c5a75cd5195768ed34d54be7/
Test: ./gradlew check
Bug:

Change-Id: I814b77d3f10754dd81f1a7e9cc5a939faf7d72f1