Add new subdirectory for the D8 version to ship in buildtools.
Adding current stable version (1.0.35).

The version used to build the platform at a given point in time
might not be the best to use for external developers. Sometimes
bugs in old VMs have to be worked around in D8. Those bugs do
not affect the platform build, but could affect other users.
Therefore, we use separate versions for building the platform
and for shipping in buildtools.

Test: make -j PRODUCT-sdk_arm64-sdk sdk_repo
Bug: 111748733

Change-Id: Ic4d1cf7ddcb009668c6ab51d6c2f02e394e4f479