Update D8 R8 master to 1.3.18-dev
Version: master 5273c3dfee1046720af425fa855316c99c86c8d2
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/5273c3dfee1046720af425fa855316c99c86c8d2/
Test: ./gradlew check
Bug:

Change-Id: I3aaf02f13e4e66979830c096d2559f584e620f15