Update D8 R8 master to 1.4.5-dev
Version: master 51cace1abdd08b320f25fb8116c2432b02f09295
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/51cace1abdd08b320f25fb8116c2432b02f09295/
Test: ./gradlew check
Bug:

Change-Id: Ie9d8bf5e507253162a169e953bafe8a07145da85