Update D8 R8 master to 1.5.1-dev
Version: master 79305435938af82784b47eb5e61543dc4a862c7f
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/79305435938af82784b47eb5e61543dc4a862c7f/
Test: ./gradlew check
Bug:

Change-Id: Id1b98f936e5571d62e86701e147652098794e0cf