Update D8 R8 master to 1.5.3-dev
Version: master 1872e9dd7d8c69ff734392165de1a04392261779
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/1872e9dd7d8c69ff734392165de1a04392261779/
Test: ./gradlew check
Bug:

Change-Id: I4bbb23aafab04d7da9d45288a71d39deb6a5374c