Update D8 R8 master to 1.5.4-dev
Version: master 96608fd7bbccdd2276cf89283d2b80cbfc4fc81c
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/96608fd7bbccdd2276cf89283d2b80cbfc4fc81c/
Test: ./gradlew check
Bug:

Change-Id: I3966302bfaba5f35a44981f54988a251ce1be8d6