Update D8 R8 master to 1.5.7-dev
Version: master faa3d5387ec055712633796ccf650e8427353edb
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/faa3d5387ec055712633796ccf650e8427353edb/
Test: ./gradlew check
Bug:

Change-Id: I3d1e1ff7fb194b90e20073247b82aefe59b737b4