Update D8 R8 master to 1.5.20-dev
Version: master a85ccd499f6d149322870a8aedddc92152071a01
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/a85ccd499f6d149322870a8aedddc92152071a01/
Test: ./gradlew check
Bug:

Change-Id: Ia6e24c7614bea29763b8b43e44dbaa8604362392