Update D8 R8 master to 1.5.21-dev
Version: master 2bc50cae723961ed287629fa26dd40af385e7dab
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/2bc50cae723961ed287629fa26dd40af385e7dab/
Test: ./gradlew check
Bug:

Change-Id: I13562c4ad245f1439ba57effebce015511509442