Update D8 R8 master to 1.6.20-dev
Version: master a8352f254885362ccd8d909d3529c60094dd896e
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/a8352f254885362ccd8d909d3529c60094dd896e/
Test: ./gradlew check
Bug: 138975804

Change-Id: I59828fc822c882e3f2714e8ab242958f44713d54