Update D8 R8 master to 1.6.21-dev
Version: master 5228491bf88bf17ad14a06abc8286fdd8f39b139
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/master/5228491bf88bf17ad14a06abc8286fdd8f39b139/
Test: ./gradlew check
Bug:

Change-Id: I9a0b3ad8926ebdb9fe24ada46778ead74328fb58