Update D8 R8 to version 1.7.1-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/1.7.1-dev/
Test: ./gradlew check

Change-Id: I3e4e8e16833412fde8492853509eb3585719aaa4