Update D8 R8 to version 1.7.6-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/1.7.6-dev/
Test: ./gradlew check
Bug:

Change-Id: I384a56bfc70eb82aa9760d4b895b7886554df98c