Update D8 R8 master to 1.7.7-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/1.7.7-dev
Test: ./gradlew check
Bug:

Change-Id: I29ff4d2c255d02aecda611fa16bf500b19c83da5