Update D8 R8 master to 1.7.12-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/1.7.12-dev
Test: ./gradlew check
Bug:

Change-Id: Ic5b18ef9a472e7fc33c736736fdefbbc616a2d6f