Update D8 R8 master to 1.7.16-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/1.7.16-dev
Test: ./gradlew check
Bug:

Change-Id: Id060f2ca42af09229335b8d31a74d0af2595baa5