Update D8 R8 master to 2.0.11-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.0.11-dev
Test: ./gradlew check
Bug:

Change-Id: Ide7f79a96ef27db10dd4191d18cb53da2b8ce9a7