Update D8 R8 master to 2.2.5-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.5-dev
Test: ./gradlew check
Bug:

Change-Id: Ib770b356c0d785a74b530380e46cdb02bfd59936