Update D8 R8 master to 2.2.6-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.6-dev
Test: ./gradlew check
Bug: 160605696

Change-Id: I732f9dbf92b7dad979a0ee0838fe374d33cdcb7f