Update D8 R8 to 2.2.13-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.13-dev
Test: ./gradlew check
Bug: 167562221

Change-Id: I33a4b8edf6043d8f3f724684771af10f67334bde