Update D8 R8 to 2.2.16-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.16-dev
Test: ./gradlew check
Bug:

Change-Id: Ic0a7a9c98b6f222cfe5b8f31b2c260cfc0c361ce