Update D8 R8 to 2.2.19-dev
This is a development snapshot, it's fine to use for studio canary build, but
not for BETA or release, for those we would need a release version of R8
binaries.
This build IS suitable for preview release but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.19-dev
Test: ./gradlew check
Bug: 169095082
Bug: 169183194

Change-Id: I398d717e0c256c0d56d1851c0dcc582d60e1819d