Update D8 R8 to 2.2.25-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.2.25-dev
Test: ./gradlew check
Bug:

Change-Id: I123f1771c6e5b386729c843c43864ab0950ed1a4