Update D8 R8 to 2.3.4-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/2.3.4-dev
Test: ./gradlew check
Bug:

Change-Id: I088d090b4468a1c53c5c7b555146d66a21369d31