Update D8 R8 to 3.0.7-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.0.7-dev
Test: ./gradlew check
Bug: 173525343

Change-Id: I3eff519c39b48e4c2d0c839bfe4bc65b4b8f2bce