Update D8 R8 to 3.0.8-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.0.8-dev
Test: ./gradlew check
Bug:

Change-Id: Ia5ad66bd5d3c44cca1d6b19e5ee1763b13f45798