Update D8 R8 to 3.0.31-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.0.31-dev
Test: ./gradlew check
Bug:

Change-Id: I57b933bd46e8a2dcdb4a6af43ace3db0bf0e712b