Update D8 R8 to 3.1.3-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.1.3-dev
Test: ./gradlew check
Bug:

Change-Id: I78920cc9a06862425e6c19eb3815cb0698b998a8