Update D8 R8 to 3.3.16-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.3.16-dev
Test: ./gradlew check
Bug:

Change-Id: Id38b530654fcc8044e1be718651e2c570f89a740