Update D8 R8 to 4.0.5-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.5-dev
Test: ./gradlew check
Bug:

Change-Id: I23efb75c820d752d2d980ea315397c13874a1112