Update D8 R8 to 4.0.18-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.18-dev
Test: ./gradlew check
Bug:

Change-Id: I0f5cac5335e2ee3353d3972d6ab903ba2c240d0a