Update D8 R8 to 8.0.13-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.0.13-dev
Test: ./gradlew check
Bug:

Change-Id: I381fee8978919174cdbe98efc85c2ff9b90f4b7a