Update D8 R8 to 8.2.7-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.2.7-dev
Test: ./gradlew check
Bug:

Change-Id: I7339d301675aa47923100c269270ab0ebbe67de8