Update D8 R8 to 8.2.20-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.2.20-dev
Test: ./gradlew check
Bug:

Change-Id: Ic838cab5aba74c7d14c32f17dd44bb552a8db42a