Update D8 R8 to 8.3.4-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.3.4-dev
Test: ./gradlew check
Bug:

Change-Id: I5f4325594f9366433fd5a488a3cd1a51b60277e1