Update D8 R8 to 8.3.7-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.3.7-dev
Test: ./gradlew check
Bug:

Change-Id: Ic168e0d8481db8cf6e0bed2ede77c47f1e5873d6