Update D8 R8 to 8.3.16-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.3.16-dev

Change-Id: I22f04f138b9d5e56e0efe1d4816ca6942b168445