Update D8 R8 to 8.3.20-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.3.20-dev

Change-Id: Ic652002d0d20ab725d6e3cf79e444d3b96b41915