Update D8 R8 to 8.4.2-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.4.2-dev

Change-Id: I1a9cc59299db734784482576d2b05ffc58bcc2d8