Update D8 R8 to 8.4.4-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.4.4-dev

Change-Id: I38e0d284c48bcf8286f5bea87f401c29a66764f8