Update D8 R8 to 8.4.7-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.4.7-dev

Change-Id: I80915c242943147d6d2449e4161f1f77718c84a0