Emit warnings for types duplicated on program and library
R8 now emits diagnostics for any type that is present on both the
program and the library/bootclasspath. For platform builds using R8
these should be treated as errors and the build configuration updated
to not have duplicates.

This CL surfaces the new diagnostics as warnings until they can
become hard errors.

Bug: 222468116
Test: m
Change-Id: I888de0a6dd799d83d0a003430a161959fb26412e