Add support for allowing library extends program in R8 full mode
This allows switching existing builds that use R8 compatibility mode to R8 full mode.

In R8 full mode, having library classes that inherit from program classes leads to a compilation error, but in R8 compatibility mode this only causes a warning. The --ignore-library-extends-program option allows lowering the error in R8 full mode to a warning to allow the build to succeed.

Bug: 215530220
Test: existing
Change-Id: Ic900dca24111ad5a27c167101e8e789066ece367